move Chat component logic to separate file
