+++ webapp/src/components/chat/Chat.tsx
@@ -1,0 +1,59 @@
+import { Subtitle1 } from '@fluentui/react-components';
+import React from 'react';
+import { AuthHelper } from '../..//libs/auth/AuthHelper';
+import { AppState, useClasses } from '../../App';
+import { UserSettingsMenu } from '../header/UserSettingsMenu';
+import { PluginGallery } from '../open-api-plugins/PluginGallery';
+import { BackendProbe, ChatView, Error, Loading } from '../views';
+
+const Chat = ({
+    classes,
+    appState,
+    setAppState,
+}: {
+    classes: ReturnType<typeof useClasses>;
+    appState: AppState;
+    setAppState: (state: AppState) => void;
+}) => {
+    const onBackendFound = React.useCallback(() => {
+        setAppState(
+            AuthHelper.isAuthAAD()
+                ? // if AAD is enabled, we need to set the active account before loading chats
+                  AppState.SettingUserInfo
+                : // otherwise, we can load chats immediately
+                  AppState.LoadingChats,
+        );
+    }, [setAppState]);
+    return (
+        <div className={classes.container}>
+            <div className={classes.header}>
+                <Subtitle1 as="h1">Chat Copilot</Subtitle1>
+                {appState > AppState.SettingUserInfo && (
+                    <div className={classes.cornerItems}>
+                        <div className={classes.cornerItems}>
+                            <PluginGallery />
+                            <UserSettingsMenu
+                                setLoadingState={() => {
+                                    setAppState(AppState.SigningOut);
+                                }}
+                            />
+                        </div>
+                    </div>
+                )}
+            </div>
+            {appState === AppState.ProbeForBackend && <BackendProbe onBackendFound={onBackendFound} />}
+            {appState === AppState.SettingUserInfo && (
+                <Loading text={'Hang tight while we fetch your information...'} />
+            )}
+            {appState === AppState.ErrorLoadingUserInfo && (
+                <Error text={'Unable to load user info. Please try signing out and signing back in.'} />
+            )}
+            {appState === AppState.ErrorLoadingChats && (
+                <Error text={'Unable to load chats. Please try refreshing the page.'} />
+            )}
+            {appState === AppState.LoadingChats && <Loading text="Loading chats..." />}
+            {appState === AppState.Chat && <ChatView />}
+        </div>
+    );
+};
+export default Chat;
